@@ -1,8 +1,5 @@
 <?php
 
-use Illuminate\Support\Facades\Facade;
-use Illuminate\Support\ServiceProvider;
-
 return [
 
 	/*
@@ -10,9 +7,9 @@ return [
 	| Application Name
 	|--------------------------------------------------------------------------
 	|
-	| This value is the name of your application. This value is used when the
+	| This value is the name of your application, which will be used when the
 	| framework needs to place the application's name in a notification or
-	| any other location as required by the application or its packages.
+	| other UI elements where an application name needs to be displayed.
 	|
 	*/
 
@@ -51,13 +48,11 @@ return [
 	|
 	| This URL is used by the console to properly generate URLs when using
 	| the Artisan command line tool. You should set this to the root of
-	| your application so that it is used when running Artisan tasks.
+	| the application so that it's available within Artisan commands.
 	|
 	*/
 
 	'url' => env('APP_URL', 'http://localhost'),
-
-	'asset_url' => env('ASSET_URL'),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -65,12 +60,12 @@ return [
 	|--------------------------------------------------------------------------
 	|
 	| Here you may specify the default timezone for your application, which
-	| will be used by the PHP date and date-time functions. We have gone
-	| ahead and set this to a sensible default for you out of the box.
+	| will be used by the PHP date and date-time functions. The timezone
+	| is set to "UTC" by default as it is suitable for most use cases.
 	|
 	*/
 
-	'timezone' => 'UTC',
+	'timezone' => env('APP_TIMEZONE', 'UTC'),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -78,53 +73,37 @@ return [
 	|--------------------------------------------------------------------------
 	|
 	| The application locale determines the default locale that will be used
-	| by the translation service provider. You are free to set this value
-	| to any of the locales which will be supported by the application.
+	| by Laravel's translation / localization methods. This option can be
+	| set to any locale for which you plan to have translation strings.
 	|
 	*/
 
-	'locale' => 'en',
+	'locale' => env('APP_LOCALE', 'en'),
 
-	/*
-	|--------------------------------------------------------------------------
-	| Application Fallback Locale
-	|--------------------------------------------------------------------------
-	|
-	| The fallback locale determines the locale to use when the current one
-	| is not available. You may change the value to correspond to any of
-	| the language folders that are provided through your application.
-	|
-	*/
+	'fallback_locale' => env('APP_FALLBACK_LOCALE', 'en'),
 
-	'fallback_locale' => 'en',
-
-	/*
-	|--------------------------------------------------------------------------
-	| Faker Locale
-	|--------------------------------------------------------------------------
-	|
-	| This locale will be used by the Faker PHP library when generating fake
-	| data for your database seeds. For example, this will be used to get
-	| localized telephone numbers, street address information and more.
-	|
-	*/
-
-	'faker_locale' => 'en_US',
+	'faker_locale' => env('APP_FAKER_LOCALE', 'en_US'),
 
 	/*
 	|--------------------------------------------------------------------------
 	| Encryption Key
 	|--------------------------------------------------------------------------
 	|
-	| This key is used by the Illuminate encrypter service and should be set
-	| to a random, 32 character string, otherwise these encrypted strings
-	| will not be safe. Please do this before deploying an application!
+	| This key is utilized by Laravel's encryption services and should be set
+	| to a random, 32 character string to ensure that all encrypted values
+	| are secure. You should do this prior to deploying the application.
 	|
 	*/
 
+	'cipher' => 'AES-256-CBC',
+
 	'key' => env('APP_KEY'),
 
-	'cipher' => 'AES-256-CBC',
+	'previous_keys' => [
+		...array_filter(
+			explode(',', env('APP_PREVIOUS_KEYS', ''))
+		),
+	],
 
 	/*
 	|--------------------------------------------------------------------------
@@ -140,50 +119,9 @@ return [
 	*/
 
 	'maintenance' => [
-		'driver' => 'file',
+		'driver' => env('APP_MAINTENANCE_DRIVER', 'file'),
+		'store' => env('APP_MAINTENANCE_STORE', 'database'),
 	],
-
-	/*
-	|--------------------------------------------------------------------------
-	| Autoloaded Service Providers
-	|--------------------------------------------------------------------------
-	|
-	| The service providers listed here will be automatically loaded on the
-	| request to your application. Feel free to add your own services to
-	| this array to grant expanded functionality to your applications.
-	|
-	*/
-
-	'providers' => ServiceProvider::defaultProviders()->merge(
-		[
-		/*
-		 * Package Service Providers...
-		 */
-
-		/*
-		 * Application Service Providers...
-		 */
-			App\Providers\AppServiceProvider::class,
-			App\Providers\AuthServiceProvider::class,
-			App\Providers\EventServiceProvider::class,
-			App\Providers\RouteServiceProvider::class,
-		]
-	)->toArray(),
-
-	/*
-	|--------------------------------------------------------------------------
-	| Class Aliases
-	|--------------------------------------------------------------------------
-	|
-	| This array of class aliases will be registered when this application
-	| is started. However, feel free to register as many as you wish as
-	| the aliases are "lazy" loaded so they don't hinder performance.
-	|
-	*/
-
-	'aliases' => Facade::defaultAliases()->merge(
-		[]
-	)->toArray(),
 
 	'frontend_url' => env('FRONTEND_URL', null),
 
