@@ -7,15 +7,15 @@ return [
 	| Authentication Defaults
 	|--------------------------------------------------------------------------
 	|
-	| This option controls the default authentication "guard" and password
-	| reset options for your application. You may change these defaults
+	| This option defines the default authentication "guard" and password
+	| reset "broker" for your application. You may change these values
 	| as required, but they're a perfect start for most applications.
 	|
 	*/
 
 	'defaults' => [
-		'guard' => 'web',
-		'passwords' => 'users',
+		'guard' => env('AUTH_GUARD', 'web'),
+		'passwords' => env('AUTH_PASSWORD_BROKER', 'users'),
 	],
 
 	/*
@@ -25,11 +25,11 @@ return [
 	|
 	| Next, you may define every authentication guard for your application.
 	| Of course, a great default configuration has been defined for you
-	| here which uses session storage and the Eloquent user provider.
+	| which utilizes session storage plus the Eloquent user provider.
 	|
-	| All authentication drivers have a user provider. This defines how the
+	| All authentication guards have a user provider, which defines how the
 	| users are actually retrieved out of your database or other storage
-	| mechanisms used by this application to persist your user's data.
+	| system used by the application. Typically, Eloquent is utilized.
 	|
 	| Supported: "session"
 	|
@@ -53,12 +53,12 @@ return [
 	| User Providers
 	|--------------------------------------------------------------------------
 	|
-	| All authentication drivers have a user provider. This defines how the
+	| All authentication guards have a user provider, which defines how the
 	| users are actually retrieved out of your database or other storage
-	| mechanisms used by this application to persist your user's data.
+	| system used by the application. Typically, Eloquent is utilized.
 	|
 	| If you have multiple user tables or models you may configure multiple
-	| sources which represent each model / table. These sources may then
+	| providers to represent the model / table. These providers may then
 	| be assigned to any extra authentication guards you have defined.
 	|
 	| Supported: "database", "eloquent"
@@ -68,7 +68,7 @@ return [
 	'providers' => [
 		'users' => [
 			'driver' => 'eloquent',
-			'model' => App\Models\User::class,
+			'model' => env('AUTH_MODEL', App\Models\User::class),
 		],
 	],
 
@@ -77,9 +77,9 @@ return [
 	| Resetting Passwords
 	|--------------------------------------------------------------------------
 	|
-	| You may specify multiple password reset configurations if you have more
-	| than one user table or model in the application and you want to have
-	| separate password reset settings based on the specific user types.
+	| These configuration options specify the behavior of Laravel's password
+	| reset functionality, including the table utilized for token storage
+	| and the user provider that is invoked to actually retrieve users.
 	|
 	| The expiry time is the number of minutes that each reset token will be
 	| considered valid. This security feature keeps tokens short-lived so
@@ -94,7 +94,7 @@ return [
 	'passwords' => [
 		'users' => [
 			'provider' => 'users',
-			'table' => 'password_resets',
+			'table' => env('AUTH_PASSWORD_RESET_TOKEN_TABLE', 'password_resets'),
 			'expire' => 60,
 			'throttle' => 60,
 		],
@@ -110,12 +110,12 @@ return [
 	|--------------------------------------------------------------------------
 	|
 	| Here you may define the amount of seconds before a password confirmation
-	| times out and the user is prompted to re-enter their password via the
+	| window expires and users are asked to re-enter their password via the
 	| confirmation screen. By default, the timeout lasts for three hours.
 	|
 	*/
 
-	'password_timeout' => 10800,
+	'password_timeout' => env('AUTH_PASSWORD_TIMEOUT', 10800),
 
 	'throttle_max_attempts_auth' => env('THROTTLE_MAX_ATTEMPTS_AUTH', 5),
 	'throttle_max_attempts_api' => env('THROTTLE_MAX_ATTEMPTS_API', 60),
