@@ -10,8 +10,9 @@ return [
 	|--------------------------------------------------------------------------
 	|
 	| Here you may specify which of the database connections below you wish
-	| to use as your default connection for all database work. Of course
-	| you may use many connections at once using the Database library.
+	| to use as your default connection for database operations. This is
+	| the connection which will be utilized unless another connection
+	| is explicitly specified when you execute a query / statement.
 	|
 	*/
 
@@ -22,14 +23,9 @@ return [
 	| Database Connections
 	|--------------------------------------------------------------------------
 	|
-	| Here are each of the database connections setup for your application.
-	| Of course, examples of configuring each database platform that is
-	| supported by Laravel is shown below to make development simple.
-	|
-	|
-	| All database work in Laravel is done through the PHP PDO facilities
-	| so make sure you have the driver for your particular database of
-	| choice installed on your machine before you begin development.
+	| Below are all of the database connections defined for your application.
+	| An example configuration is provided for each database system which
+	| is supported by Laravel. You're free to add / remove connections.
 	|
 	*/
 
@@ -37,43 +33,64 @@ return [
 
 		'sqlite' => [
 			'driver' => 'sqlite',
-			'url' => env('DATABASE_URL'),
+			'url' => env('DB_URL'),
 			'database' => env('DB_DATABASE', database_path('database.sqlite')),
 			'prefix' => '',
 			'foreign_key_constraints' => env('DB_FOREIGN_KEYS', true),
+			'busy_timeout' => null,
+			'journal_mode' => null,
+			'synchronous' => null,
 		],
 
 		'mysql' => [
 			'driver' => 'mysql',
-			'url' => env('DATABASE_URL'),
+			'url' => env('DB_URL'),
 			'host' => env('DB_HOST', '127.0.0.1'),
 			'port' => env('DB_PORT', '3306'),
-			'database' => env('DB_DATABASE', 'forge'),
-			'username' => env('DB_USERNAME', 'forge'),
+			'database' => env('DB_DATABASE', 'laravel'),
+			'username' => env('DB_USERNAME', 'root'),
 			'password' => env('DB_PASSWORD', ''),
 			'unix_socket' => env('DB_SOCKET', ''),
-			'charset' => 'utf8mb4',
-			'collation' => 'utf8mb4_unicode_ci',
+			'charset' => env('DB_CHARSET', 'utf8mb4'),
+			'collation' => env('DB_COLLATION', 'utf8mb4_unicode_ci'),
 			'prefix' => '',
 			'prefix_indexes' => true,
 			'strict' => true,
 			'engine' => null,
-			'options' => extension_loaded('pdo_mysql') ? array_filter(
-				[
-					PDO::MYSQL_ATTR_SSL_CA => env('MYSQL_ATTR_SSL_CA'),
-				]
-			) : [],
+			'options' => extension_loaded('pdo_mysql') ? array_filter([
+				PDO::MYSQL_ATTR_SSL_CA => env('MYSQL_ATTR_SSL_CA'),
+			]) : [],
+		],
+
+		'mariadb' => [
+			'driver' => 'mariadb',
+			'url' => env('DB_URL'),
+			'host' => env('DB_HOST', '127.0.0.1'),
+			'port' => env('DB_PORT', '3306'),
+			'database' => env('DB_DATABASE', 'laravel'),
+			'username' => env('DB_USERNAME', 'root'),
+			'password' => env('DB_PASSWORD', ''),
+			'unix_socket' => env('DB_SOCKET', ''),
+			'charset' => env('DB_CHARSET', 'utf8mb4'),
+			'collation' => env('DB_COLLATION', 'utf8mb4_unicode_ci'),
+			'prefix' => '',
+			'prefix_indexes' => true,
+			'strict' => true,
+			'engine' => null,
+			'options' => extension_loaded('pdo_mysql') ? array_filter([
+				PDO::MYSQL_ATTR_SSL_CA => env('MYSQL_ATTR_SSL_CA'),
+			]) : [],
 		],
 
 		'pgsql' => [
 			'driver' => 'pgsql',
-			'url' => env('DATABASE_URL'),
+			'url' => env('DB_URL'),
 			'host' => env('DB_HOST', '127.0.0.1'),
 			'port' => env('DB_PORT', '5432'),
-			'database' => env('DB_DATABASE', 'forge'),
-			'username' => env('DB_USERNAME', 'forge'),
+			'database' => env('DB_DATABASE', 'laravel'),
+			'username' => env('DB_USERNAME', 'root'),
 			'password' => env('DB_PASSWORD', ''),
-			'charset' => 'utf8',
+			'charset' => env('DB_CHARSET', 'utf8'),
 			'prefix' => '',
 			'prefix_indexes' => true,
 			'search_path' => 'public',
@@ -82,13 +99,13 @@ return [
 
 		'sqlsrv' => [
 			'driver' => 'sqlsrv',
-			'url' => env('DATABASE_URL'),
+			'url' => env('DB_URL'),
 			'host' => env('DB_HOST', 'localhost'),
 			'port' => env('DB_PORT', '1433'),
-			'database' => env('DB_DATABASE', 'forge'),
-			'username' => env('DB_USERNAME', 'forge'),
+			'database' => env('DB_DATABASE', 'laravel'),
+			'username' => env('DB_USERNAME', 'root'),
 			'password' => env('DB_PASSWORD', ''),
-			'charset' => 'utf8',
+			'charset' => env('DB_CHARSET', 'utf8'),
 			'prefix' => '',
 			'prefix_indexes' => true,
 		],
@@ -102,11 +119,14 @@ return [
 	|
 	| This table keeps track of all the migrations that have already run for
 	| your application. Using this information, we can determine which of
-	| the migrations on disk haven't actually been run in the database.
+	| the migrations on disk haven't actually been run on the database.
 	|
 	*/
 
-	'migrations' => 'migrations',
+	'migrations' => [
+		'table' => 'migrations',
+		'update_date_on_publish' => true,
+	],
 
 	/*
 	|--------------------------------------------------------------------------
@@ -115,7 +135,7 @@ return [
 	|
 	| Redis is an open source, fast, and advanced key-value store that also
 	| provides a richer body of commands than a typical key-value system
-	| such as APC or Memcached. Laravel makes it easy to dig right in.
+	| such as Memcached. You may define your connection settings here.
 	|
 	*/
 
