@@ -9,7 +9,7 @@ return [
 	|
 	| Here you may specify the default filesystem disk that should be used
 	| by the framework. The "local" disk, as well as a variety of cloud
-	| based disks are available to your application. Just store away!
+	| based disks are available to your application for file storage.
 	|
 	*/
 
@@ -20,11 +20,11 @@ return [
 	| Filesystem Disks
 	|--------------------------------------------------------------------------
 	|
-	| Here you may configure as many filesystem "disks" as you wish, and you
-	| may even configure multiple disks of the same driver. Defaults have
-	| been set up for each driver as an example of the required values.
+	| Below you may configure as many filesystem disks as necessary, and you
+	| may even configure multiple disks for the same driver. Examples for
+	| most supported storage drivers are configured here for reference.
 	|
-	| Supported Drivers: "local", "ftp", "sftp", "s3"
+	| Supported drivers: "local", "ftp", "sftp", "s3"
 	|
 	*/
 
@@ -32,7 +32,8 @@ return [
 
 		'local' => [
 			'driver' => 'local',
-			'root' => storage_path('app'),
+			'root' => storage_path('app/private'),
+			'serve' => true,
 			'throw' => false,
 		],
 
