@@ -12,9 +12,9 @@ return [
 	| Default Log Channel
 	|--------------------------------------------------------------------------
 	|
-	| This option defines the default log channel that gets used when writing
-	| messages to the logs. The name specified in this option should match
-	| one of the channels defined in the "channels" configuration array.
+	| This option defines the default log channel that is utilized to write
+	| messages to your logs. The value provided here should match one of
+	| the channels present in the list of "channels" configured below.
 	|
 	*/
 
@@ -33,7 +33,7 @@ return [
 
 	'deprecations' => [
 		'channel' => env('LOG_DEPRECATIONS_CHANNEL', 'null'),
-		'trace' => false,
+		'trace' => env('LOG_DEPRECATIONS_TRACE', false),
 	],
 
 	/*
@@ -41,20 +41,20 @@ return [
 	| Log Channels
 	|--------------------------------------------------------------------------
 	|
-	| Here you may configure the log channels for your application. Out of
-	| the box, Laravel uses the Monolog PHP logging library. This gives
-	| you a variety of powerful log handlers / formatters to utilize.
+	| Here you may configure the log channels for your application. Laravel
+	| utilizes the Monolog PHP logging library, which includes a variety
+	| of powerful log handlers and formatters that you're free to use.
 	|
-	| Available Drivers: "single", "daily", "slack", "syslog",
-	|                    "errorlog", "monolog",
-	|                    "custom", "stack"
+	| Available drivers: "single", "daily", "slack", "syslog",
+	|                    "errorlog", "monolog", "custom", "stack"
 	|
 	*/
 
 	'channels' => [
+
 		'stack' => [
 			'driver' => 'stack',
-			'channels' => ['single'],
+			'channels' => explode(',', env('LOG_STACK', 'single')),
 			'ignore_exceptions' => false,
 		],
 
@@ -69,15 +69,15 @@ return [
 			'driver' => 'daily',
 			'path' => storage_path('logs/laravel.log'),
 			'level' => env('LOG_LEVEL', 'debug'),
-			'days' => 14,
+			'days' => env('LOG_DAILY_DAYS', 14),
 			'replace_placeholders' => true,
 		],
 
 		'slack' => [
 			'driver' => 'slack',
 			'url' => env('LOG_SLACK_WEBHOOK_URL'),
-			'username' => 'Laravel Log',
-			'emoji' => ':boom:',
+			'username' => env('LOG_SLACK_USERNAME', 'Laravel Log'),
+			'emoji' => env('LOG_SLACK_EMOJI', ':boom:'),
 			'level' => env('LOG_LEVEL', 'critical'),
 			'replace_placeholders' => true,
 		],
@@ -108,7 +108,7 @@ return [
 		'syslog' => [
 			'driver' => 'syslog',
 			'level' => env('LOG_LEVEL', 'debug'),
-			'facility' => LOG_USER,
+			'facility' => env('LOG_SYSLOG_FACILITY', LOG_USER),
 			'replace_placeholders' => true,
 		],
 
@@ -140,6 +140,7 @@ return [
 			'path' => storage_path('logs/database.log'),
 			'level' => env('LOG_LEVEL', 'debug'),
 		],
+
 	],
 
 	'database' => env('LOG_DATABASE_QUERIES', false),
