@@ -7,13 +7,14 @@ return [
 	| Default Mailer
 	|--------------------------------------------------------------------------
 	|
-	| This option controls the default mailer that is used to send any email
-	| messages sent by your application. Alternative mailers may be setup
-	| and used as needed; however, this mailer will be used by default.
+	| This option controls the default mailer that is used to send all email
+	| messages unless another mailer is explicitly specified when sending
+	| the message. All additional mailers can be configured within the
+	| "mailers" array. Examples of each type of mailer are provided.
 	|
 	*/
 
-	'default' => env('MAIL_MAILER', 'smtp'),
+	'default' => env('MAIL_MAILER', 'log'),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -24,38 +25,40 @@ return [
 	| their respective settings. Several examples have been configured for
 	| you and you are free to add your own as your application requires.
 	|
-	| Laravel supports a variety of mail "transport" drivers to be used while
-	| sending an e-mail. You will specify which one you are using for your
-	| mailers below. You are free to add additional mailers as required.
+	| Laravel supports a variety of mail "transport" drivers that can be used
+	| when delivering an email. You may specify which one you're using for
+	| your mailers below. You may also add additional mailers if needed.
 	|
 	| Supported: "smtp", "sendmail", "mailgun", "ses", "ses-v2",
-	|            "postmark", "log", "array", "failover"
+	|            "postmark", "resend", "log", "array",
+	|            "failover", "roundrobin"
 	|
 	*/
 
 	'mailers' => [
+
 		'smtp' => [
 			'transport' => 'smtp',
+			'scheme' => env('MAIL_SCHEME'),
 			'url' => env('MAIL_URL'),
-			'host' => env('MAIL_HOST', 'smtp.mailgun.org'),
-			'port' => env('MAIL_PORT', 587),
-			'encryption' => env('MAIL_ENCRYPTION', 'tls'),
+			'host' => env('MAIL_HOST', '127.0.0.1'),
+			'port' => env('MAIL_PORT', 2525),
 			'username' => env('MAIL_USERNAME'),
 			'password' => env('MAIL_PASSWORD'),
 			'timeout' => null,
-			'local_domain' => env('MAIL_EHLO_DOMAIN'),
+			'local_domain' => env('MAIL_EHLO_DOMAIN', parse_url(env('APP_URL', 'http://localhost'), PHP_URL_HOST)),
 		],
 
 		'ses' => [
 			'transport' => 'ses',
 		],
 
-		'mailgun' => [
-			'transport' => 'mailgun',
-		],
-
 		'postmark' => [
 			'transport' => 'postmark',
+		],
+
+		'resend' => [
+			'transport' => 'resend',
 		],
 
 		'sendmail' => [
@@ -79,6 +82,15 @@ return [
 				'log',
 			],
 		],
+
+		'roundrobin' => [
+			'transport' => 'roundrobin',
+			'mailers' => [
+				'ses',
+				'postmark',
+			],
+		],
+
 	],
 
 	/*
@@ -86,9 +98,9 @@ return [
 	| Global "From" Address
 	|--------------------------------------------------------------------------
 	|
-	| You may wish for all e-mails sent by your application to be sent from
-	| the same address. Here, you may specify a name and address that is
-	| used globally for all e-mails that are sent by your application.
+	| You may wish for all emails sent by your application to be sent from
+	| the same address. Here you may specify a name and address that is
+	| used globally for all emails that are sent by your application.
 	|
 	*/
 
