@@ -7,22 +7,22 @@ return [
 	| Default Queue Connection Name
 	|--------------------------------------------------------------------------
 	|
-	| Laravel's queue API supports an assortment of back-ends via a single
-	| API, giving you convenient access to each back-end using the same
-	| syntax for every one. Here you may define a default connection.
+	| Laravel's queue supports a variety of backends via a single, unified
+	| API, giving you convenient access to each backend using identical
+	| syntax for each. The default queue connection is defined below.
 	|
 	*/
 
-	'default' => env('QUEUE_CONNECTION', 'sync'),
+	'default' => env('QUEUE_CONNECTION', 'database'),
 
 	/*
 	|--------------------------------------------------------------------------
 	| Queue Connections
 	|--------------------------------------------------------------------------
 	|
-	| Here you may configure the connection information for each server that
-	| is used by your application. A default configuration has been added
-	| for each back-end shipped with Laravel. You are free to add more.
+	| Here you may configure the connection options for every queue backend
+	| used by your application. An example configuration is provided for
+	| each backend supported by Laravel. You're also free to add more.
 	|
 	| Drivers: "sync", "database", "beanstalkd", "sqs", "redis", "null"
 	|
@@ -36,17 +36,18 @@ return [
 
 		'database' => [
 			'driver' => 'database',
-			'table' => 'jobs',
-			'queue' => 'default',
-			'retry_after' => 90,
+			'connection' => env('DB_QUEUE_CONNECTION'),
+			'table' => env('DB_QUEUE_TABLE', 'jobs'),
+			'queue' => env('DB_QUEUE', 'default'),
+			'retry_after' => (int) env('DB_QUEUE_RETRY_AFTER', 90),
 			'after_commit' => false,
 		],
 
 		'beanstalkd' => [
 			'driver' => 'beanstalkd',
-			'host' => 'localhost',
-			'queue' => 'default',
-			'retry_after' => 90,
+			'host' => env('BEANSTALKD_QUEUE_HOST', 'localhost'),
+			'queue' => env('BEANSTALKD_QUEUE', 'default'),
+			'retry_after' => (int) env('BEANSTALKD_QUEUE_RETRY_AFTER', 90),
 			'block_for' => 0,
 			'after_commit' => false,
 		],
@@ -64,9 +65,9 @@ return [
 
 		'redis' => [
 			'driver' => 'redis',
-			'connection' => 'default',
+			'connection' => env('REDIS_QUEUE_CONNECTION', 'default'),
 			'queue' => env('REDIS_QUEUE', 'default'),
-			'retry_after' => 90,
+			'retry_after' => (int) env('REDIS_QUEUE_RETRY_AFTER', 90),
 			'block_for' => null,
 			'after_commit' => false,
 		],
@@ -95,8 +96,10 @@ return [
 	|--------------------------------------------------------------------------
 	|
 	| These options configure the behavior of failed queue job logging so you
-	| can control which database and table are used to store the jobs that
-	| have failed. You may change them to any database / table you wish.
+	| can control how and where failed jobs are stored. Laravel ships with
+	| support for storing failed jobs in a simple file or in a database.
+	|
+	| Supported drivers: "database-uuids", "dynamodb", "file", "null"
 	|
 	*/
 
