@@ -82,8 +82,8 @@ return [
 
 	'middleware' => [
 		'authenticate_session' => Laravel\Sanctum\Http\Middleware\AuthenticateSession::class,
-		'encrypt_cookies' => App\Http\Middleware\EncryptCookies::class,
-		'verify_csrf_token' => App\Http\Middleware\VerifyCsrfToken::class,
+		'encrypt_cookies' => Illuminate\Cookie\Middleware\EncryptCookies::class,
+		'verify_csrf_token' => Illuminate\Foundation\Http\Middleware\ValidateCsrfToken::class,
 	],
 
 ];
