@@ -14,13 +14,6 @@ return [
 	|
 	*/
 
-	'mailgun' => [
-		'domain' => env('MAILGUN_DOMAIN'),
-		'secret' => env('MAILGUN_SECRET'),
-		'endpoint' => env('MAILGUN_ENDPOINT', 'api.mailgun.net'),
-		'scheme' => 'https',
-	],
-
 	'postmark' => [
 		'token' => env('POSTMARK_TOKEN'),
 	],
@@ -29,6 +22,17 @@ return [
 		'key' => env('AWS_ACCESS_KEY_ID'),
 		'secret' => env('AWS_SECRET_ACCESS_KEY'),
 		'region' => env('AWS_DEFAULT_REGION', 'us-east-1'),
+	],
+
+	'resend' => [
+		'key' => env('RESEND_KEY'),
+	],
+
+	'slack' => [
+		'notifications' => [
+			'bot_user_oauth_token' => env('SLACK_BOT_USER_OAUTH_TOKEN'),
+			'channel' => env('SLACK_BOT_USER_DEFAULT_CHANNEL'),
+		],
 	],
 
 ];
