@@ -9,9 +9,9 @@ return [
 	| Default Session Driver
 	|--------------------------------------------------------------------------
 	|
-	| This option controls the default session "driver" that will be used on
-	| requests. By default, we will use the lightweight native driver but
-	| you may specify any of the other wonderful drivers provided here.
+	| This option determines the default session driver that is utilized for
+	| incoming requests. Laravel supports a variety of storage options to
+	| persist session data. Database storage is a great default choice.
 	|
 	| Supported: "file", "cookie", "database", "apc",
 	|            "memcached", "redis", "dynamodb", "array"
@@ -27,13 +27,14 @@ return [
 	|
 	| Here you may specify the number of minutes that you wish the session
 	| to be allowed to remain idle before it expires. If you want them
-	| to immediately expire on the browser closing, set that option.
+	| to expire immediately when the browser is closed then you may
+	| indicate that via the expire_on_close configuration option.
 	|
 	*/
 
 	'lifetime' => env('SESSION_LIFETIME', 120),
 
-	'expire_on_close' => false,
+	'expire_on_close' => env('SESSION_EXPIRE_ON_CLOSE', false),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -41,21 +42,21 @@ return [
 	|--------------------------------------------------------------------------
 	|
 	| This option allows you to easily specify that all of your session data
-	| should be encrypted before it is stored. All encryption will be run
-	| automatically by Laravel and you can use the Session like normal.
+	| should be encrypted before it's stored. All encryption is performed
+	| automatically by Laravel and you may use the session like normal.
 	|
 	*/
 
-	'encrypt' => false,
+	'encrypt' => env('SESSION_ENCRYPT', false),
 
 	/*
 	|--------------------------------------------------------------------------
 	| Session File Location
 	|--------------------------------------------------------------------------
 	|
-	| When using the native session driver, we need a location where session
-	| files may be stored. A default has been set for you but a different
-	| location may be specified. This is only needed for file sessions.
+	| When utilizing the "file" session driver, the session files are placed
+	| on disk. The default storage location is defined here; however, you
+	| are free to provide another location where they should be stored.
 	|
 	*/
 
@@ -79,22 +80,22 @@ return [
 	| Session Database Table
 	|--------------------------------------------------------------------------
 	|
-	| When using the "database" session driver, you may specify the table we
-	| should use to manage the sessions. Of course, a sensible default is
-	| provided for you; however, you are free to change this as needed.
+	| When using the "database" session driver, you may specify the table to
+	| be used to store sessions. Of course, a sensible default is defined
+	| for you; however, you're welcome to change this to another table.
 	|
 	*/
 
-	'table' => 'sessions',
+	'table' => env('SESSION_TABLE', 'sessions'),
 
 	/*
 	|--------------------------------------------------------------------------
 	| Session Cache Store
 	|--------------------------------------------------------------------------
 	|
-	| While using one of the framework's cache driven session backends you may
-	| list a cache store that should be used for these sessions. This value
-	| must match with one of the application's configured cache "stores".
+	| When using one of the framework's cache driven session backends, you may
+	| define the cache store which should be used to store the session data
+	| between requests. This must match one of your defined cache stores.
 	|
 	| Affects: "apc", "dynamodb", "memcached", "redis"
 	|
@@ -120,9 +121,9 @@ return [
 	| Session Cookie Name
 	|--------------------------------------------------------------------------
 	|
-	| Here you may change the name of the cookie used to identify a session
-	| instance by ID. The name specified here will get used every time a
-	| new session cookie is created by the framework for every driver.
+	| Here you may change the name of the session cookie that is created by
+	| the framework. Typically, you should not need to change this value
+	| since doing so does not grant a meaningful security improvement.
 	|
 	*/
 
@@ -138,20 +139,20 @@ return [
 	|
 	| The session cookie path determines the path for which the cookie will
 	| be regarded as available. Typically, this will be the root path of
-	| your application but you are free to change this when necessary.
+	| your application, but you're free to change this when necessary.
 	|
 	*/
 
-	'path' => '/',
+	'path' => env('SESSION_PATH', '/'),
 
 	/*
 	|--------------------------------------------------------------------------
 	| Session Cookie Domain
 	|--------------------------------------------------------------------------
 	|
-	| Here you may change the domain of the cookie used to identify a session
-	| in your application. This will determine which domains the cookie is
-	| available to in your application. A sensible default has been set.
+	| This value determines the domain and subdomains the session cookie is
+	| available to. By default, the cookie will be available to the root
+	| domain and all subdomains. Typically, this shouldn't be changed.
 	|
 	*/
 
@@ -177,11 +178,11 @@ return [
 	|
 	| Setting this value to true will prevent JavaScript from accessing the
 	| value of the cookie and the cookie will only be accessible through
-	| the HTTP protocol. You are free to modify this option if needed.
+	| the HTTP protocol. It's unlikely you should disable this option.
 	|
 	*/
 
-	'http_only' => true,
+	'http_only' => env('SESSION_HTTP_ONLY', true),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -190,13 +191,15 @@ return [
 	|
 	| This option determines how your cookies behave when cross-site requests
 	| take place, and can be used to mitigate CSRF attacks. By default, we
-	| will set this value to "lax" since this is a secure default value.
+	| will set this value to "lax" to permit secure cross-site requests.
+	|
+	| See: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie#samesitesamesite-value
 	|
 	| Supported: "lax", "strict", "none", null
 	|
 	*/
 
-	'same_site' => 'lax',
+	'same_site' => env('SESSION_SAME_SITE', 'lax'),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -209,6 +212,6 @@ return [
 	|
 	*/
 
-	'partitioned' => false,
+	'partitioned' => env('SESSION_PARTITIONED_COOKIE', false),
 
 ];
